refactor(pagination): return i64 from Pager::limit and Pager::offset
Switch `Pager::limit()` from `usize` to `i64` and `Pager::offset()` from
`u64` to `i64` (saturating) so callsites that bind them as sqlx
LIMIT/OFFSET parameters don't need to repeat the cast. `Pager::size`
becomes `u16` internally to match its input and keep every conversion
infallible.

`dbmcp_sql::pagination::with_limit_offset` takes `i64`/`i64` to match.
Postgres `list_tables` drops its local `i64::try_from(...).unwrap_or(i64::MAX)`
pair and binds `pager.limit()` / `pager.offset()` directly.

No behavioral change: i64's Display is identical to usize/u64 for
non-negative values, so all string-interpolated callsites emit the same
SQL. Integration matrix passes across all four backends.

+++ crates/sql/src/pagination.rs
@@ -12,7 +12,7 @@
 /// new line so a trailing `-- line comment` in the caller's SQL cannot
 /// swallow the wrap's own tokens.
 #[must_use]
-pub fn with_limit_offset(sql: &str, limit: usize, offset: u64) -> String {
+pub fn with_limit_offset(sql: &str, limit: i64, offset: i64) -> String {
     let trimmed = sql.trim_end();
     let inner = trimmed.strip_suffix(';').unwrap_or(trimmed).trim_end();
     format!("SELECT * FROM ({inner}\n) AS paginated LIMIT {limit} OFFSET {offset}")
